refactor: reduce number of rebuilds in user cluster
